requirements.txt
pytesseract
Pillow
matplotlib
fpdf
PyPDF2
pandas
